Add bindings for match context
This binds 2 vars during the evaluation of targets.

`mach.core/*target-ctx*` this is bound to a string that contains the
whole string that matched this target e.g. a glob fo* would match
"foo.xml", this var would contain "foo.xml"

`mach.core/*matcher-ctx*` this is bound to the object that was used to
match this target.  e.g. for `#mach/glob "fo*"` the underlying Glob
object will be bound to this var.