mach test: use mach classpath so andare is added
Running `mach test` was failing with errors about core.async not being
found. That's because andare jar was not being added to the classpath as
it's done in bin/mach. We can simply take lumo's classpath that includes
that jar, because we are in the context of the lumo process started from
our bin/mach command